[REF] core: move DummyRLock to tests
The lock is not needed for normal execution. We are moving it to tests.

This PR also fixes `TestAuthLDAP` which is not working (using inexisting
functions and classmethod on setup).
