+++ odoo/modules/registry/__init__.py
@@ -1,3 +1,3 @@
 # ruff: noqa: F401
 # Exposed here so that exist code is unaffected.
-from odoo.orm.registry import DummyRLock, Registry, _REGISTRY_CACHES, _CACHES_BY_KEY
+from odoo.orm.registry import Registry, _REGISTRY_CACHES, _CACHES_BY_KEY
